+++ src/main/java/micycle/pgs/commons/RepulsionCirclePack.java
@@ -23,8 +23,9 @@ import processing.core.PVector;
  * a toroid. Each circle's centre is constrained to lie within the rectangle but
  * its edges are allowed to extend outside.
  * <p>
- * This forms a Java implementation of the algorithm from the <i>packcircles</i>
- * R package.
+ * This Java code is based on an implementation of the algorithm from the
+ * <i>packcircles</i> R package, but adds grid-based indexing to speed up the
+ * packing convergence.
  * 
  * @author Michael Carleton
  *
@@ -34,12 +35,8 @@ public class RepulsionCirclePack {
 	// based on
 	// https://github.com/mbedward/packcircles/blob/main/src/packcircles.cpp
 
-	/*-
-	 * TODO:
-	 * R-tree (or similar) to do intersection checks,
-	 * PShape version: place circles with weight=1 on perimeter, then pack circles inside shape
-	 * https://web.archive.org/web/20080517011454/http://en.wiki.mcneel.com/default.aspx/McNeel/2DCirclePacking
-	 */
+	// consider
+	// https://web.archive.org/web/20080517011454/http://en.wiki.mcneel.com/default.aspx/McNeel/2DCirclePacking
 
 	private static boolean USE_GRID = true;
 	private static double almostZero_TOL = 1e-2;
@@ -52,16 +49,16 @@ public class RepulsionCirclePack {
 	 */
 	private List<Double> weights;
 
-	double cellSize;
+	private double cellSize;
+	private final float xmin;
+	private final float xmax;
+	private final float ymin;
+	private final float ymax;
+	private final boolean wrap;
 
-	double xmin;
-	double xmax;
-	double ymin;
-	double ymax;
-
-	int gridWidth, gridHeight;
-	List<List<PVector>> grid;
-	Map<PVector, float[]> updates;
+	private int gridWidth, gridHeight;
+	private List<List<PVector>> grid;
+	private Map<PVector, float[]> updates;
 
 	/**
 	 * 
@@ -73,17 +70,15 @@ public class RepulsionCirclePack {
 	 * @param wrap
 	 */
 	public RepulsionCirclePack(List<PVector> xyr, double xmin, double xmax, double ymin, double ymax, final boolean wrap) {
-		this.xmin = xmin;
-		this.xmax = xmax;
-		this.ymin = ymin;
-		this.ymax = ymax;
-		this.circles = xyr.stream().map(v -> v.copy()).collect(Collectors.toList()); // TODO hilbert sort?
-//		this.circles = PGS_PointSet.hilbertSort(circles);
-//		circles.sort((a,b) -> Float.compare(a.x+a.y, b.x+b.y));
+		this.circles = xyr.stream().map(PVector::copy).collect(Collectors.toList());
+		this.xmin = (float) xmin;
+		this.xmax = (float) xmax;
+		this.ymin = (float) ymin;
+		this.ymax = (float) ymax;
+		this.wrap = wrap;
 
 		if (USE_GRID) {
-
-			float maxR = circles.stream().max((a, b) -> Float.compare(a.z, b.z)).orElse(new PVector(0,0,1)).z;
+			float maxR = circles.stream().max((a, b) -> Float.compare(a.z, b.z)).orElse(new PVector(0, 0, 2)).z;
 			cellSize = (maxR * 2 + almostZero_TOL);
 			gridWidth = (int) Math.ceil((xmax - xmin) / cellSize) + 4;
 			gridHeight = (int) Math.ceil((ymax - ymin) / cellSize) + 4;
@@ -93,39 +88,10 @@ public class RepulsionCirclePack {
 				grid.add(new ArrayList<>());
 			}
 			circles.forEach(this::addToGrid);
-//		System.out.println(grid == null);
 			updates = new HashMap<>();
 		}
 
-		iterateLayout(xmin, xmax, ymin, ymax, wrap);
-//		System.out.println(minR);
-//		System.out.println(gridHeight*gridWidth);
-	}
-
-	private void addToGrid(PVector p) {
-		grid.get(index(p.x, p.y)).add(p);
-	}
-
-	private int index(double x, double y) {
-//offset grid cell by 2 in each direction to account for border
-		int gx = (int) Math.floor((x - xmin) / cellSize) + 2;
-		int gy = (int) Math.floor((y - ymin) / cellSize) + 2;
-		return gy * gridWidth + gx;
-	}
-
-	private int indexY(double x, double y) {
-		return (int) Math.floor((y - ymin) / cellSize) + 2;
-	}
-
-	private int indexX(double x, double y) {
-		return (int) Math.floor((x - xmin) / cellSize) + 2;
-	}
-
-	private List<PVector> atIndex(int indexX, int indexY) {
-		return grid.get(indexY * gridWidth + indexX);
-	}
-
-	void buildGrid() {
+		iterateLayout();
 	}
 
 	public List<PVector> getPacking() {
@@ -140,15 +106,11 @@ public class RepulsionCirclePack {
 	 * @param weights vector of double values between 0 and 1, used as
 	 *                multiplicative weights for the distance a circle will move
 	 *                with pair repulsion
-	 * @param xmin    bounds min X
-	 * @param xmax    bounds max X
-	 * @param ymin    bounds min Y
-	 * @param ymax    bounds max Y
 	 * @param maxiter maximum number of iterations
-	 * @param wrap    allow coordinate wrapping across opposite bounds
+	 * 
 	 * @return the number of iterations performed
 	 */
-	private int iterateLayout(double xmin, double xmax, double ymin, double ymax, final boolean wrap) {
+	private int iterateLayout() {
 		final int N = circles.size();
 		int iter = 0;
 
@@ -158,22 +120,21 @@ public class RepulsionCirclePack {
 			if (!USE_GRID) {
 				for (int i = 0; i < N - 1; ++i) {
 					for (int j = i + 1; j < N; ++j) {
-						if (doRepulsion(circles.get(i), circles.get(j), xmin, xmax, ymin, ymax, wrap, false)) {
+						if (doRepulsion(circles.get(i), circles.get(j), false)) {
 							moved = true;
 						}
 					}
 				}
 			} else {
 				for (PVector circle : circles) {
-//				int index = index(circle.x, circle.y);
-					int indexX = indexX(circle.x, circle.y);
-					int indexY = indexY(circle.x, circle.y);
+					int indexX = indexX(circle.x);
+					int indexY = indexY(circle.y);
 					// MARK WHETHER we've already done the repulsion in the loop
 					// since we do everything twice
 					for (int i = indexX - 1; i < indexX + 1; i++) {
 						for (int j = indexY - 1; j < indexY + 2; j++) { // NOTE +2
 							for (PVector neighbor : atIndex(i, j)) {
-								if (neighbor != circle && doRepulsion(circle, neighbor, xmin, xmax, ymin, ymax, wrap, true)) {
+								if (neighbor != circle && doRepulsion(circle, neighbor, true)) {
 									moved = true;
 								}
 							}
@@ -183,25 +144,13 @@ public class RepulsionCirclePack {
 				}
 			}
 
-			if (!moved) {
+			if (!moved) { // didn't move any -- no circles overlapping
 				break;
 			}
 
 		}
 
-//		System.out.println(iter);
 		return iter;
-	}
-
-	void updateGrid() {
-		updates.forEach((c, pos) -> {
-			int index = index(c.x, c.y); // original
-			grid.get(index).remove(c);
-			c.x = pos[0];
-			c.y = pos[1];
-			addToGrid(c); // reinsert at correct position
-		});
-		updates.clear();
 	}
 
 	/**
@@ -209,39 +158,28 @@ public class RepulsionCirclePack {
 	 * distance moved by each circle is proportional to the radius of the other to
 	 * give some semblance of intertia.
 	 * 
-	 * @param c0   index of first circle
-	 * @param c1   index of second circle
-	 * @param xmin bounds min X
-	 * @param xmax bounds max X
-	 * @param ymin bounds min Y
-	 * @param ymax bounds max Y
-	 * @param wrap allow coordinate wrapping across opposite bounds
+	 * @param c0 first circle
+	 * @param c1 second circle
 	 */
-	private boolean doRepulsion(PVector c0, PVector c1, double xmin, double xmax, double ymin, double ymax, final boolean wrap,
-			boolean useGrid) {
-
-		// TODO every iteration, place into triangulation, repulse only connected
-		// circles; find based on pinwheel ?
-
-		// if both weights are zero, return zero to indicate
-		// no movement
+	private boolean doRepulsion(final PVector c0, final PVector c1, final boolean useGrid) {
+		// if both weights are zero, return zero to indicate no movement
 //		if (almostZero(weights.get(c0)) && almostZero(weights.get(c1))) {
 //			return 0;
 //		}
 
-		double dx = c1.x - c0.x;
-		double dy = c1.y - c0.y;
-		double d = Math.sqrt(dx * dx + dy * dy);
-		double r = c1.z + c0.z;
-		double p;
-		double w0;
-		double w1;
+		float dx = c1.x - c0.x;
+		float dy = c1.y - c0.y;
+		float d = (float) Math.sqrt(dx * dx + dy * dy);
+		float r = c1.z + c0.z;
+		float p;
+		float w0;
+		float w1;
 
 		if (gtZero(r - d)) {
 			if (almostZero(d)) {
 				// The two centres are coincident or almost so.
 				// Arbitrarily move along x-axis
-				p = 1.0;
+				p = 1.0f;
 				dx = r - d;
 			} else {
 				p = (r - d) / d;
@@ -254,12 +192,13 @@ public class RepulsionCirclePack {
 
 			float c1XNew = ordinate(c1.x + p * dx * w1, xmin, xmax, wrap);
 			float c1YNew = ordinate(c1.y + p * dy * w1, ymin, ymax, wrap);
-			int c1Index = index(c1.x, c1.y); // original
+			int c1Index = index(c1.x, c1.y);
 			if (useGrid && c1Index != index(c1XNew, c1YNew)) {
-				// TODO ideally update position immediately
+				/*
+				 * Can't change since we're currently iterating one of the grid cells
+				 */
 				updates.put(c1, new float[] { c1XNew, c1YNew });
-			} else {
-				// update coords but not grid reference
+			} else { // update coords but not grid reference
 				c1.x = c1XNew;
 				c1.y = c1YNew;
 			}
@@ -267,11 +206,10 @@ public class RepulsionCirclePack {
 			float c0XNew = ordinate(c0.x - p * dx * w0, xmin, xmax, wrap);
 			float c0YNew = ordinate(c0.y - p * dy * w0, ymin, ymax, wrap);
 
-			int c0Index = index(c0.x, c0.y); // original
+			int c0Index = index(c0.x, c0.y);
 			if (useGrid && c0Index != index(c0XNew, c0YNew)) {
 				updates.put(c0, new float[] { c0XNew, c0YNew });
-			} else {
-				// update coords but not grid reference
+			} else { // update coords but not grid reference
 				c0.x = c0XNew;
 				c0.y = c0YNew;
 			}
@@ -282,24 +220,45 @@ public class RepulsionCirclePack {
 		return false;
 	}
 
-	class Circle {
-		float x, y;
-		float r;
-		int index; // used to reference to the object (since x,y can change) and update its grid
-					// reference
-		int provisionalGridIndex;
-		
-		@Override
-		public int hashCode() {
-			return index;
-		}
+	private void addToGrid(PVector p) {
+		grid.get(index(p.x, p.y)).add(p);
 	}
 
-	private static boolean almostZero(double x) {
+	private void updateGrid() {
+		updates.forEach((c, pos) -> {
+			int index = index(c.x, c.y); // original
+			grid.get(index).remove(c);
+			c.x = pos[0];
+			c.y = pos[1];
+			addToGrid(c); // reinsert at correct position
+		});
+		updates.clear();
+	}
+
+	private int index(double x, double y) {
+		// offset grid cell by 2 in each direction to account for border
+		int gx = (int) Math.floor((x - xmin) / cellSize) + 2;
+		int gy = (int) Math.floor((y - ymin) / cellSize) + 2;
+		return gy * gridWidth + gx;
+	}
+
+	private int indexY(double y) {
+		return (int) Math.floor((y - ymin) / cellSize) + 2;
+	}
+
+	private int indexX(double x) {
+		return (int) Math.floor((x - xmin) / cellSize) + 2;
+	}
+
+	private List<PVector> atIndex(int indexX, int indexY) {
+		return grid.get(indexY * gridWidth + indexX);
+	}
+
+	private static boolean almostZero(float x) {
 		return Math.abs(x) < almostZero_TOL;
 	}
 
-	private static boolean gtZero(double x) {
+	private static boolean gtZero(float x) {
 		return !almostZero(x) && (x > 0.0);
 	}
 
@@ -307,23 +266,22 @@ public class RepulsionCirclePack {
 	 * Adjust an X or Y ordinate to the given bounds by either wrapping (if `wrap`
 	 * is true) or clamping (if `wrap` is false).
 	 */
-
-	private static float ordinate(double x, double lo, double hi, final boolean wrap) {
+	private static float ordinate(float x, float lo, float hi, final boolean wrap) {
 		if (wrap) {
-			return (float) wrapOrdinate(x, lo, hi);
+			return wrapOrdinate(x, lo, hi);
 		} else {
-			return (float) Math.max(lo, Math.min(hi, x));
+			return Math.max(lo, Math.min(hi, x));
 		}
 	}
 
 	/**
-	 * Map an X or Y ordinate to the toroidal interval [lo, hi).
+	 * Map an X or Y ordinate to the toroidal interval [lo, hi].
 	 *
 	 * x - X or Y ordinate to be adjusted lo - lower coordinate bound hi - upper
 	 * coordinate bound
 	 */
-	private static double wrapOrdinate(double x, double lo, double hi) {
-		double w = hi - lo;
+	private static float wrapOrdinate(float x, float lo, float hi) {
+		float w = hi - lo;
 		while (x < lo) {
 			x += w;
 		}